LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-to-combination-of-two-elements-2-expected.txt


FAIL Snap to points of combinations of two different elements assert_equals: expected 250 but got 200
